requirements.txt
arrow==1.3.0
asgiref==3.9.1
blessed==1.21.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
ctransformers==0.2.27
Django==5.2.4
django-picklefield==3.3
django-q==1.3.9
django-q2==1.8.0
djlint==1.36.4
dnspython==2.7.0
EditorConfig==0.17.1
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
jsbeautifier==1.15.4
json5==0.12.0
mongoengine==0.29.1
packaging==25.0
pathspec==0.12.1
py-cpuinfo==9.0.0
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==3.5.3
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
django-allauth>=0.65.0
